# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Art Gallery Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PaintingSize {
  name: string;
  width: number;
  height: number;
  unit: string;
}

export interface Painting {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  imageUrl: string;
  sizes: PaintingSize[];
  inStock: boolean;
  featured: boolean;
  createdAt: string;
}

export interface PaintingInput {
  title: string;
  description?: string;
  price: number;
  imageUrl: string;
  sizes: PaintingSize[];
  inStock?: boolean;
  featured?: boolean;
}

export interface PaintingUpdate {
  title?: string;
  description?: string;
  price?: number;
  imageUrl?: string;
  sizes?: PaintingSize[];
  inStock?: boolean;
  featured?: boolean;
}

export interface OrderItem {
  paintingId: number;
  paintingTitle: string;
  size: string;
  price: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  address: string;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  totalPrice: number;
  status: OrderStatus;
  createdAt: string;
}

export interface OrderInput {
  customerName: string;
  phone: string;
  address: string;
  notes?: string;
  items: OrderItem[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface OrderStats {
  total: number;
  pending: number;
  confirmed: number;
  shipped: number;
  delivered: number;
  cancelled: number;
  totalRevenue: number;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  username: string;
  loggedIn: boolean;
}