# lib/api-zod/src/generated/types/paintingSize.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Art Gallery Store API
 * OpenAPI spec version: 0.1.0
 */

export interface PaintingSize {
  name: string;
  width: number;
  height: number;
  unit: string;
}