# lib/api-zod/src/generated/types/paintingUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Art Gallery Store API
 * OpenAPI spec version: 0.1.0
 */
import type { PaintingSize } from './paintingSize';

export interface PaintingUpdate {
  title?: string;
  description?: string;
  price?: number;
  imageUrl?: string;
  sizes?: PaintingSize[];
  inStock?: boolean;
  featured?: boolean;
}